+++ include/mata/util.hh
@@ -28,8 +28,6 @@
 #include <stack>
 #include <unordered_map>
 #include <vector>
-#include <mata/ord-vector.hh>
-#include <mata/number-predicate.hh>
 
 /// macro for debug outputs
 #define PRINT_VERBOSE_LVL(lvl, title, x) {\
@@ -60,19 +58,6 @@ extern unsigned LOG_VERBOSITY;
 namespace Util
 {
 
-template <typename Number>
-bool are_disjoint(NumberPredicate<Number> lhs, NumberPredicate<Number> rhs) {
-    return lhs.are_disjoint(rhs);
-}
-
-template <typename Number>
-bool are_disjoint(Mata::Util::OrdVector<Number> lhs, NumberPredicate<Number> rhs) {
-    for (auto q: lhs)
-        if (rhs[q])
-            return false;
-    return true;
-}
-
 /** Are two sets disjoint? */
 template <class T>
 bool are_disjoint(const std::set<T>& lhs, const std::set<T>& rhs)
@@ -87,22 +72,6 @@ bool are_disjoint(const std::set<T>& lhs, const std::set<T>& rhs)
 	}
 
 	return true;
-} // }}}
-
-
-template <class T>
-bool are_disjoint(const Util::OrdVector<T>& lhs, const Util::OrdVector<T>& rhs)
-{ // {{{
-    auto itLhs = lhs.begin();
-    auto itRhs = rhs.begin();
-    while (itLhs != lhs.end() && itRhs != rhs.end())
-    {
-        if (*itLhs == *itRhs) { return false; }
-        else if (*itLhs < *itRhs) { ++itLhs; }
-        else {++itRhs; }
-    }
-
-    return true;
 } // }}}
 
 /** Is there an element in a container? */
@@ -426,7 +395,7 @@ std::string to_string(const A& value)
 
 namespace Mata
 {
-namespace util
+namespace Util
 {
 
 // Taken from
