+++ src/strings/tests-nfa-noodlification.cc
@@ -10,7 +10,7 @@
 
 using namespace Mata::Nfa;
 using namespace Mata::Strings;
-using namespace Mata::util;
+using namespace Mata::Util;
 using namespace Mata::Parser;
 using namespace Mata::RE2Parser;
